+++ src/project/controllers/project.controller.ts
@@ -14,10 +14,10 @@ import { JwtAuthGuard } from 'src/auth/guards/jwt.guard';
 import { RolesGuard } from 'src/auth/guards/roles.guard';
 import { Roles } from 'src/shared/decorators/roles.decorator';
 import { Role } from 'src/shared/enums/role.enum';
-import { CreateProjectRequestDto } from './dto/request/create-project.dto';
-import { ModifyProjectRequestDto } from './dto/request/modify-project.dto';
-import { ProjectParamsDto } from './dto/request/project-params.dto';
-import { ProjectService } from './project.service';
+import { CreateProjectRequestDto } from '../dto/request/create-project.dto';
+import { ModifyProjectRequestDto } from '../dto/request/modify-project.dto';
+import { ProjectParamsDto } from '../dto/request/project-params.dto';
+import { ProjectService } from '../services/project.service';
 
 @Controller('project')
 export class ProjectController {
